Remove stern warning from README
I'd like to suggest the strongly worded warning be removed from the
README. This library is solving a real problem that Django applications
can frequently run into. It is better for there to exist a community
driven accepted approach rather than each project re-implementing it
with a unique set of bugs and shortcomings. This warning has been used
in projects to reject this package as a dependency. I think this library
can be brought to a production ready state such that it can be used as a
dependency of other projects.